made float32

+++ WeightBias.py
@@ -6,11 +6,11 @@ class WeightBias:
     def __init__(self,name,in_len,out_len):
         self.name = name
 
-        rand_weight_vals = np.random.randn(in_len*out_len)/in_len**(0.5**0.5)
+        rand_weight_vals = np.random.randn(in_len*out_len).astype('float32')/in_len**(0.5**0.5)
         rand_weight = np.reshape(rand_weight_vals,(out_len,in_len))
         self.W = theano.shared(rand_weight,name=self.weight_name())
 
-        bias_init = np.zeros(out_len)
+        bias_init = np.zeros(out_len,dtype='float32')
         self.b = theano.shared(bias_init,name=self.bias_name())#allows nice addition
 
     def calc_output(self,in_vec):
